+++ pci/Cargo.toml
@@ -10,7 +10,7 @@ byteorder = "1.3.4"
 devices = { path = "../devices" }
 kvm-bindings = { git = "https://github.com/cloud-hypervisor/kvm-bindings", branch = "ch" }
 kvm-ioctls = { git = "https://github.com/cloud-hypervisor/kvm-ioctls", branch = "ch" }
-vfio = { path = "../vfio" }
+vfio-ioctls = { path = "../vfio-ioctls" }
 vmm-sys-util = ">=0.3.1"
 libc = "0.2.71"
 log = "0.4.8"
