+++ vmm/Cargo.toml
@@ -7,7 +7,7 @@ edition = "2018"
 [features]
 default = []
 acpi = ["acpi_tables","devices/acpi", "arch/acpi"]
-pci_support = ["pci", "vfio", "vm-virtio/pci_support"]
+pci_support = ["pci", "vfio-ioctls", "vm-virtio/pci_support"]
 mmio_support = ["vm-virtio/mmio_support"]
 cmos = ["devices/cmos"]
 fwdebug = ["devices/fwdebug"]
@@ -34,7 +34,7 @@ serde = {version = ">=1.0.27", features = ["rc"] }
 serde_derive = ">=1.0.27"
 serde_json = ">=1.0.9"
 url = "2.1.1"
-vfio = { path = "../vfio", optional = true }
+vfio-ioctls = { path = "../vfio-ioctls", optional = true }
 vm-allocator = { path = "../vm-allocator" }
 vm-device = { path = "../vm-device" }
 vm-memory = { version = "0.2.1", features = ["backend-mmap", "backend-atomic"] }
